chore: Add error handling and container for MongoDB
