+++ .env
@@ -1,1 +1,11 @@
-MONGODB_URI="mongodb+srv://starwars:b2wdigital@cluster0.v6mbo.mongodb.net/starwars?retryWrites=true&w=majority"
+# Application
+HOST=localhost
+PORT=3333
+
+
+# MongoDB
+DB_HOST=localhost
+DB_PORT=27017
+DB_NAME=starwars
+DB_USER=starwars
+DB_PASS=b2wdigital
