+++ .env.example
@@ -1,0 +1,10 @@
+# Application
+HOST=
+PORT=
+
+# MongoDB
+DB_HOST=
+DB_PORT=
+DB_NAME=
+DB_USER=
+DB_PASS=
