+++ Makefile
@@ -1,0 +1,16 @@
+include .env
+
+.PHONY: up
+
+up:
+	docker-compose up -d
+
+.PHONY: down
+
+down:
+	docker-compose down
+
+.PHONY: logs
+
+logs:
+	docker-compose logs -f
