+++ docker-compose.yaml
@@ -1,0 +1,12 @@
+version: "3.2"
+services:
+  db:
+    image: mongo:4.2
+    container_name: mongo
+    restart: always
+    ports:
+      - ${DB_PORT}:27017
+    environment:
+      MONGO_INITDB_DATABASE: ${DB_NAME}
+      MONGO_INITDB_ROOT_USERNAME: ${DB_USER}
+      MONGO_INITDB_ROOT_PASSWORD: ${DB_PASS}
