+++ framework/storage/init.go
@@ -1,0 +1,5 @@
+package storage
+
+func init() {
+	connectDB()
+}
