+++ framework/storage/mongodb.go
@@ -1,0 +1,38 @@
+package storage
+
+import (
+	"context"
+	"log"
+	"os"
+	"time"
+
+	"github.com/joho/godotenv"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+// DB holds MongoDB connection
+var DB mongo.Database
+
+// ConnectDB returns Client and Context to connect to database
+func connectDB() {
+
+	err := godotenv.Load()
+	if err != nil {
+		log.Fatalf("Error loading .env file")
+		panic(err)
+	}
+
+	uri := "mongodb://" + os.Getenv("DB_USER") + ":" + os.Getenv("DB_PASS") + "@" + os.Getenv("DB_HOST") + ":" + os.Getenv("DB_PORT")
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	client, err := mongo.Connect(ctx, options.Client().ApplyURI(uri))
+	if err != nil {
+		log.Fatalf("Error connecting to MongoDB: %v", err)
+		panic(err)
+	}
+
+	DB = *client.Database(os.Getenv("DB_NAME"))
+}
